notebooks_and_code/contrastive_learning/requirements.txt
# Core dependencies for ConPLex-style contrastive learning
torch>=1.9.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0
PyYAML>=5.4.0

# Optional dependencies
wandb>=0.12.0              # For experiment tracking
tensorboard>=2.7.0         # Alternative logging
jupyterlab>=3.0.0         # For notebook development

# Bio-related dependencies (if using real molecular data)
rdkit>=2021.09.1          # Molecular representation
biopython>=1.79           # Protein sequence analysis

# Existing EnzyFind dependencies
unimol_tools              # Molecular representation
fair-esm                  # Protein embeddings
py-xgboost               # Baseline comparison